refactor prompt state utilities into package
